Move route-helpers to modules
